Create Automated Email Report with PDF ZIP and Excel.js
Creates an email that is generated automatically with Sheets and PDFs in a folder attached as an Excel and ZIP file respectively. Removes the need for user to manually download, upload, write the email and update the dates. Creates a copy of the ZIP file sent within the Drive folder where the PDFs were extracted.